+++ Source/Modules/ISHServer/Initialize-ISHIIS.ps1
@@ -21,7 +21,7 @@ function Initialize-ISHIIS
     
     begin 
     {
-        . $PSScriptRoot\Test-RunningAsElevated.ps1
+        . $PSScriptRoot\Private\Test-RunningAsElevated.ps1
         Test-RunningAsElevated -StopCallerPSCmdlet $PSCmdlet
     }