+++ Source/Modules/ISHServer/Initialize-ISHRegionalDefault.ps1
@@ -21,7 +21,7 @@ function Initialize-ISHRegionalDefault
     
     begin 
     {
-        . $PSScriptRoot\Test-RunningAsElevated.ps1
+        . $PSScriptRoot\Private\Test-RunningAsElevated.ps1
         Test-RunningAsElevated -StopCallerPSCmdlet $PSCmdlet
     }