+++ Source/Modules/ISHServer/Initialize-ISHUser.ps1
@@ -24,7 +24,7 @@ function Initialize-ISHUser
     
     begin 
     {
-        . $PSScriptRoot\Test-RunningAsElevated.ps1
+        . $PSScriptRoot\Private\Test-RunningAsElevated.ps1
         Test-RunningAsElevated -StopCallerPSCmdlet $PSCmdlet
     }