+++ Source/Modules/ISHServer/Install-ISHToolJavaHelp.ps1
@@ -23,7 +23,7 @@ function Install-ISHToolJavaHelp
     
     begin 
     {
-        . $PSScriptRoot\Test-RunningAsElevated.ps1
+        . $PSScriptRoot\Private\Test-RunningAsElevated.ps1
         Test-RunningAsElevated -StopCallerPSCmdlet $PSCmdlet
     }