+++ Source/Modules/ISHServer/Private/Get-ISHFTPItem.ps1
@@ -1,0 +1,52 @@
+<#
+# Copyright (c) 2014 All Rights Reserved by the SDL Group.
+# 
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+#     http://www.apache.org/licenses/LICENSE-2.0
+# 
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#>
+
+function Get-ISHFTPItem
+{
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory=$true)]
+        [string]$FTPHost,
+        [Parameter(Mandatory=$true)]
+        [pscredential]$Credential,
+        [Parameter(Mandatory=$true)]
+        [string[]]$Path,
+        [Parameter(Mandatory=$true)]
+        [string]$LocalPath
+    )
+    
+    begin 
+    {
+        Import-Module PSFTP -ErrorAction Stop
+        Set-FTPConnection -Server $FTPHost -Credentials $Credential -UseBinary -KeepAlive -UsePassive | Out-Null
+    }
+
+    process
+    {
+        $Path | ForEach-Object {
+            Write-Debug "Path=$_"
+            Get-FTPItem -Path $_ -LocalPath $LocalPath -Overwrite
+            Write-Verbose "Downloaded $_ to $LocalPath"
+            $filePath=Join-Path $LocalPath ($_.Substring($_.LastIndexOf('/')+1))
+            Get-Item -Path $filePath
+        }
+    }
+
+    end
+    {
+
+    }
+}