+++ Source/Modules/ISHServer/Set-ISHFirewallHTTPS.ps1
@@ -21,7 +21,7 @@ function Set-ISHFirewallHTTPS
     
     begin 
     {
-        . $PSScriptRoot\Test-RunningAsElevated.ps1
+        . $PSScriptRoot\Private\Test-RunningAsElevated.ps1
         Test-RunningAsElevated -StopCallerPSCmdlet $PSCmdlet
     }